Move MVT specific code from service-mvt-wfs to service-mvt
